+++ tests/testthat/test-test_get_all_rejection_rate.R
@@ -1,0 +1,46 @@
+library(testthat)
+
+test_that("Get all rejection rates", {
+
+model_simple_med <-
+"
+m ~ x
+y ~ m + x
+"
+
+model_simple_med_es <- c("y ~ m" = "l",
+                         "m ~ x" = "m",
+                         "y ~ x" = "n")
+
+sim_only <- power4test(nrep = 4,
+                       model = model_simple_med,
+                       pop_es = model_simple_med_es,
+                       n = 100,
+                       R = 50,
+                       ci_type = "boot",
+                       fit_model_args = list(fit_function = "lm"),
+                       do_the_test = FALSE,
+                       iseed = 1234)
+
+test_out <- power4test(object = sim_only,
+                       test_fun = test_indirect_effect,
+                       test_args = list(x = "x",
+                                        m = "m",
+                                        y = "y",
+                                        boot_ci = TRUE,
+                                        mc_ci = FALSE))
+test_out <- power4test(object = test_out,
+                       test_fun = test_parameters)
+test_out <- power4test(object = test_out,
+                       test_fun = test_parameters,
+                       test_args = list(op = "~"))
+test_out <- power4test(object = test_out,
+                       test_fun = test_indirect_effect,
+                       test_args = list(x = "x",
+                                        y = "y",
+                                        boot_ci = TRUE,
+                                        mc_ci = FALSE))
+
+expect_no_error(get_rejection_rates(test_out))
+
+})